+++ chrome-remote-cache/src/index.ts
@@ -1,5 +1,6 @@
 export { ChromeRemoteCache } from './ChromeRemoteCache';
 export { CacheStat } from './CacheStat';
 export { CacheManager } from './CacheManager';
+export { dropQueryParam, formatSize } from './CacheUtils';
 
 export { ChromeRemoteCache as default } from './ChromeRemoteCache';