+++ chrome-remote-cache/CHANGELOG.md
@@ -1,2 +1,8 @@
+# changelog
+
+## v 0.0.2 (2022-07-28)
+* add methods to ChromeRemoteCache
+* add dropQueryParam, formatSize exports
+
 ## v 0.0.1 (2022-07-28)
-initial release
+* initial release